requirements.txt
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.3
curl_cffi==0.15.0
deprecation==2.1.0
frozendict==2.4.7
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.15
markdown-it-py==4.2.0
mdurl==0.1.2
multidict==6.7.1
multitasking==0.0.13
numpy==2.4.5
packaging==26.2
pandas==3.0.3
peewee==4.0.5
platformdirs==4.9.6
postgrest==2.25.1
propcache==0.5.2
protobuf==7.34.1
pycparser==3.0
pydantic==2.13.4
pydantic_core==2.46.4
Pygments==2.20.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.2
realtime==2.25.1
requests==2.34.2
rich==15.0.0
six==1.17.0
soupsieve==2.8.3
storage3==2.25.1
StrEnum==0.4.15
supabase==2.25.1
supabase-auth==2.25.1
supabase-functions==2.25.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.7.0
websockets==15.0.1
yarl==1.23.0
yfinance==1.3.0
